Edits to API import functiong

+++ wrangle.py
@@ -65,35 +65,36 @@ def data_clean(df):
     df['observation_time']= pd.to_datetime(df['observation_time'])
     return df
 
-def get_climacell_data():
-    # obtain new data
-    df = api_get()
-    # clean new data
-    df = data_clean(df)
-    # obtain previous data
-    df_old = pd.read_csv('climate_data.csv', index_col=0)
-    # add new data on to previous data
-    merge = [df_old, df]
-    df = pd.concat(merge)
-    df.reset_index(drop = True)
-    # drop column added by concat
-    #df.drop(columns =['Unnamed: 0'], inplace = True)
-    # convert observation to date time
-    df['observation_time'] =  pd.to_datetime(df['observation_time'])
-    # Filtering out any repeated time observations 
-    df.sort_values("observation_time", inplace = True) 
-    # dropping any duplicte values 
-    df.drop_duplicates(subset ="observation_time", 
+def get_climacell_data(cached = False):
+    '''
+    This function reads in climate data from API if cached == False 
+    or if cached == True reads in zillow df from a csv file, returns df
+    '''
+    if cached or os.path.isfile('climate_data.csv') == False:
+        # obtain new data
+        df = api_get()
+        # clean new data
+        df = data_clean(df)
+        # obtain previous data
+        df_old = pd.read_csv('climate_data.csv', index_col=0)
+        # add new data on to previous data
+        merge = [df_old, df]
+        df = pd.concat(merge)
+        df.reset_index(drop = True)
+        # drop column added by concat
+        #df.drop(columns =['Unnamed: 0'], inplace = True)
+        # convert observation to date time
+        df['observation_time'] =  pd.to_datetime(df['observation_time'])
+        # Filtering out any repeated time observations 
+        df.sort_values("observation_time", inplace = True) 
+        # dropping any duplicte values 
+        df.drop_duplicates(subset ="observation_time", 
                      keep = False, inplace = True) 
-    # add daily column to data frame
-    #df['day'] = df.observation_time.dt.day
-    # add month column to data frame
-    # df['month'] = df.observation_time.dt.month
-    # set date as index
-    #df = df.set_index('observation_time').sort_index()
-    # write new data to csv
-    df.to_csv('climate_data.csv')
-    # return new data frame
+        # write new data to csv
+        df.to_csv('climate_data.csv')
+        # return new data frame
+    else:
+        df = pd.read_csv('climate_data.csv', index_col=0)
     return df
 
 def formatt_data(df):
